feat: only allow currentUser's to react
